Visitor class takes compiler and options
